Fixes #21950: Remove ncf scripts
