--- tests/quick/test_ncf_api_assets/technique_metadata_test_content.cf
@@ -1,16 +1,0 @@
-# @name ncf technique method argument escape test
-# @description This is a bundle to test ncf's Python lib
-# @version 0.1
-
-bundle agent content_escaping_test
-{
-  vars:
-    "resources_dir" string => "${this.promise_dir}/resources";
-  methods:
-    "Install a package with correct version_context_${report_data.directive_id}_0" usebundle => _method_reporting_context("Install a package with correct version", "apache2");
-    "Install a package with correct version" usebundle => package_install_version("apache2", "2.2.11"),
-      ifvarclass => concat("any");
-    "Edit conf file_context_${report_data.directive_id}_1" usebundle => _method_reporting_context("Edit conf file", "/etc/httpd/conf/httpd.conf");
-    "Edit conf file" usebundle => file_replace_lines("/etc/httpd/conf/httpd.conf", "ErrorLog \"/var/log/httpd/error_log\"", "ErrorLog \"/projet/logs/httpd/error_log\""),
-      ifvarclass => concat("redhat");
-}
